pyproject.toml
[project]
name = "sermon-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "dynaconf>=3.2.12",
    "fastapi>=0.131.0",
    "fastapi-cache2>=0.2.2",
    "google-generativeai>=0.8.6",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-jose>=3.5.0",
    "redis>=7.3.0",
    "reportlab>=4.4.10",
    "rich>=14.3.3",
    "sentry-sdk>=2.54.0",
    "slowapi>=0.1.9",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]
